+++ lib/glimmer/dsl/opal/display_expression.rb
@@ -20,13 +20,20 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 require 'glimmer/dsl/static_expression'
-require 'glimmer/dsl/opal/exec_expression'
+require 'glimmer/dsl/parent_expression'
+require 'glimmer/dsl/top_level_expression'
+require 'glimmer/swt/display_proxy'
 
 module Glimmer
   module DSL
     module Opal
-      class AsyncExecExpression < StaticExpression
-        include ExecExpression
+      class DisplayExpression < StaticExpression
+        include TopLevelExpression
+        include ParentExpression
+
+        def interpret(parent, keyword, *args, &block)
+          Glimmer::SWT::DisplayProxy.instance(*args)
+        end
       end
     end
   end